old_requirements.txt
absl-py
aiohttp
aiosignal
astor==0.8.1
astunparse==1.6.3
async-timeout
asynctest==0.13.0
attrs 
blinker==1.4
brotlipy==0.7.0
cachetools
certifi 
cffi 
charset-normalizer 
click
cloudpickle 
colorama 
cryptography 
cycler==0.11.0
Farama-Notifications==0.0.4
flit_core 
fonttools==4.38.0
frozenlist
future 
gast==0.3.3
google-auth 
google-auth-oauthlib 
google-pasta
grpcio 
#gym
gymnasium
h5py
idna 
importlib-metadata 
jax-jumpy
joblib==1.3.2
Keras-Applications
Keras-Preprocessing
kiwisolver==1.4.5
Markdown 
matplotlib
mkl-fft
mkl-random 
mkl-service==2.4.0
multidict
numpy 
oauthlib
opt-einsum 
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
protobuf==3.20.3
PuLP==2.8.0
pyasn1 
pyasn1-modules==0.2.8
pycparser
pyglet
PyJWT 
pyOpenSSL 
pyparsing==3.1.2
pyreadline==2.1
PySocks 
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests
requests-oauthlib==1.3.0
rsa 
scipy 
Shimmy==1.1.0
six 
stable-baselines3==2.0.0
tensorboard 
termcolor
torch==2.1.1
typing_extensions 
urllib3 
Werkzeug 
win-inet-pton
wincertstore==0.2
wrapt 
yarl
zipp 
